ratatui_termwiz

Struct TermwizBackend

pub struct TermwizBackend { /* private fields */ }

A Backend implementation that uses Termwiz to render to the terminal.

The TermwizBackend struct is a wrapper around a BufferedTerminal, which is used to send commands to the terminal. It provides methods for drawing content, manipulating the cursor, and clearing the terminal screen.

Most applications should not call the methods on TermwizBackend directly, but will instead use the Terminal struct, which provides a more ergonomic interface.

This backend automatically enables raw mode and switches to the alternate screen when it is created using the TermwizBackend::new method (and disables raw mode and returns to the main screen when dropped). Use the TermwizBackend::with_buffered_terminal to create a new instance with a custom BufferedTerminal if this is not desired.

Example

use ratatui::Terminal;
use ratatui::backend::TermwizBackend;

let backend = TermwizBackend::new()?;
let mut terminal = Terminal::new(backend)?;

terminal.clear()?;
terminal.draw(|frame| {
    // -- snip --
})?;

See the the Examples directory for more examples. See the backend module documentation for more details on raw mode and alternate screen.

Implementations

impl TermwizBackend

pub fn new() -> Result<Self, Box<dyn Error>>

Creates a new Termwiz backend instance.

The backend will automatically enable raw mode and enter the alternate screen.

Errors

Returns an error if unable to do any of the following:

query the terminal capabilities.

enter raw mode.

enter the alternate screen.

create the system or buffered terminal.

Example

use ratatui::backend::TermwizBackend;

let backend = TermwizBackend::new()?;

pub const fn with_buffered_terminal( instance: BufferedTerminal<SystemTerminal>, ) -> Self

Creates a new Termwiz backend instance with the given buffered terminal.

pub const fn buffered_terminal(&self) -> &BufferedTerminal<SystemTerminal>

Returns a reference to the buffered terminal used by the backend.

pub const fn buffered_terminal_mut( &mut self, ) -> &mut BufferedTerminal<SystemTerminal>

Returns a mutable reference to the buffered terminal used by the backend.