ratatui_termwiz

Trait FromTermwiz

pub trait FromTermwiz<T> {
    // Required method
    fn from_termwiz(termwiz: T) -> Self;
}

A trait for converting types from Termwiz to Ratatui.

This trait replaces the From trait for converting types from Termwiz to Ratatui. It is necessary because the From trait is not implemented for types defined in external crates.

Required Methods

fn from_termwiz(termwiz: T) -> Self

Converts the given Termwiz type to the Ratatui type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.